requirements.txt
Flask==2.3.3
python-barcode==0.15.1
Pillow==10.0.1
gunicorn==21.2.0
qrcode[pil]==7.4.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7